+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/permission/PermissionRequest.kt
@@ -74,6 +74,7 @@ sealed class Permission {
     data class ContentVideoOther(override val id: String? = "", override val desc: String? = "") : Permission()
     data class ContentAutoPlayAudible(override val id: String? = "", override val desc: String? = "") : Permission()
     data class ContentAutoPlayInaudible(override val id: String? = "", override val desc: String? = "") : Permission()
+    data class ContentPersistentStorage(override val id: String? = "", override val desc: String? = "") : Permission()
 
     data class AppCamera(override val id: String? = "", override val desc: String? = "") : Permission()
     data class AppAudio(override val id: String? = "", override val desc: String? = "") : Permission()
@@ -82,7 +83,7 @@ sealed class Permission {
 
     data class Generic(override val id: String?, override val desc: String? = "") : Permission()
 
-    // Removed in GeckoView 68.0:
+    // Removed in GeckoView 68.0: //TO REMOVE
     data class ContentVideoApplication(override val id: String? = "", override val desc: String? = "") : Permission()
     data class ContentVideoBrowser(override val id: String? = "", override val desc: String? = "") : Permission()
     data class ContentVideoWindow(override val id: String? = "", override val desc: String? = "") : Permission()